refactor tests/support into logical submodules
Split monolithic tests/support/mod.rs (~640 lines) into:
- fixtures.rs: Fixture struct, constructors, matrix functions
- git.rs: init_git_repo, run_git, copy_tree, write_worktree_diff, setup_fixture_worktree
- runner.rs: consolidated run_covgate + run_covgate_raw (env_clear semantics)
- parity.rs: OverallTotals, MetricFixtureCase, write_* helpers

mod.rs is now a pure re-export hub. Deleted assert_fixture_has_no_branch_coverage
(unused), run_covgate_with_coverage, run_covgate_with_env, and the local
run_covgate_raw_with_path in cli_interface.rs. Callers updated across all test files.
config_discovery.rs and config_auto_base.rs now import run_git from support.
