+++ tests/cli_interface.rs
@@ -1,24 +1,14 @@
 mod support;
 
-use std::{fs, path::PathBuf, process::Output};
+use std::{fs, path::PathBuf};
 
 use tempfile::tempdir;
 
 use crate::support::{
-    copy_tree, init_git_repo, run_covgate, run_covgate_raw, run_covgate_with_coverage,
-    run_covgate_with_env, run_git, rust_basic_fail_fixture, rust_basic_pass_fixture,
-    setup_fixture_worktree, vitest_path_scoped_gates_fixture, write_absolute_path_coverage_fixture,
-    write_worktree_diff,
+    copy_tree, init_git_repo, run_covgate, run_covgate_raw, run_git, rust_basic_fail_fixture,
+    rust_basic_pass_fixture, setup_fixture_worktree, vitest_path_scoped_gates_fixture,
+    write_absolute_path_coverage_fixture, write_worktree_diff,
 };
-
-fn run_covgate_raw_with_path(worktree: &std::path::Path, path: &str, args: &[String]) -> Output {
-    let binary = env!("CARGO_BIN_EXE_covgate");
-    let mut command = std::process::Command::new(binary);
-    command.args(args);
-    command.current_dir(worktree);
-    command.env("PATH", path);
-    command.output().expect("covgate should run")
-}
 
 fn setup_path_scoped_fixture() -> (tempfile::TempDir, PathBuf, PathBuf) {
     let fixture = vitest_path_scoped_gates_fixture();
@@ -35,7 +25,7 @@ fn record_base_noops_when_standard_base_ref_is_available() {
     let worktree = setup_fixture_worktree(temp.path(), fixture);
     run_git(&worktree, &["branch", "-M", "main"]);
 
-    let output = run_covgate_raw(&worktree, &["record-base".to_string()]);
+    let output = run_covgate_raw(&worktree, &["record-base".to_string()], &[]);
     assert_eq!(output.status.code(), Some(0));
     let stdout = String::from_utf8(output.stdout).expect("stdout should be utf8");
     assert!(
@@ -62,7 +52,7 @@ fn record_base_creates_worktree_ref_in_constrained_repo() {
     let worktree = setup_fixture_worktree(temp.path(), fixture);
     run_git(&worktree, &["branch", "-M", "task/record-base"]);
 
-    let output = run_covgate_raw(&worktree, &["record-base".to_string()]);
+    let output = run_covgate_raw(&worktree, &["record-base".to_string()], &[]);
     assert_eq!(output.status.code(), Some(0));
     let stdout = String::from_utf8(output.stdout).expect("stdout should be utf8");
     assert!(stdout.contains("Recorded base commit"), "stdout={stdout}");
@@ -86,7 +76,7 @@ fn record_base_does_not_break_git_ref_enumeration() {
     let worktree = setup_fixture_worktree(temp.path(), fixture);
     run_git(&worktree, &["branch", "-M", "task/record-base"]);
 
-    let output = run_covgate_raw(&worktree, &["record-base".to_string()]);
+    let output = run_covgate_raw(&worktree, &["record-base".to_string()], &[]);
     assert_eq!(output.status.code(), Some(0));
 
     let show_ref = std::process::Command::new("git")
@@ -105,7 +95,7 @@ fn record_base_does_not_break_git_ref_enumeration() {
 fn record_base_fails_outside_git_repo() {
     let temp = tempdir().expect("tempdir should exist");
 
-    let output = run_covgate_raw(temp.path(), &["record-base".to_string()]);
+    let output = run_covgate_raw(temp.path(), &["record-base".to_string()], &[]);
     assert_eq!(output.status.code(), Some(1));
     let stderr = String::from_utf8(output.stderr).expect("stderr should be utf8");
     assert!(
@@ -118,7 +108,7 @@ fn record_base_fails_outside_git_repo() {
 fn record_base_fails_fast_when_git_is_missing() {
     let temp = tempdir().expect("tempdir should exist");
 
-    let output = run_covgate_raw_with_path(temp.path(), "", &["record-base".to_string()]);
+    let output = run_covgate_raw(temp.path(), &["record-base".to_string()], &[("PATH", "")]);
 
     assert_eq!(output.status.code(), Some(1));
     let stderr = String::from_utf8(output.stderr).expect("stderr should be utf8");
@@ -132,7 +122,7 @@ fn record_base_fails_fast_when_git_is_missing() {
 fn missing_check_subcommand_is_reported_as_clap_usage_error() {
     let temp = tempdir().expect("tempdir should exist");
 
-    let output = run_covgate_raw(temp.path(), &[]);
+    let output = run_covgate_raw(temp.path(), &[], &[]);
     assert_eq!(output.status.code(), Some(2));
     let stderr = String::from_utf8(output.stderr).expect("stderr should be utf8");
     assert!(
@@ -146,7 +136,7 @@ fn missing_check_subcommand_is_reported_as_clap_usage_error() {
 fn missing_check_coverage_report_is_reported_as_clap_usage_error() {
     let temp = tempdir().expect("tempdir should exist");
 
-    let output = run_covgate_raw(temp.path(), &["check".to_string()]);
+    let output = run_covgate_raw(temp.path(), &["check".to_string()], &[]);
     assert_eq!(output.status.code(), Some(2));
     let stderr = String::from_utf8(output.stderr).expect("stderr should be utf8");
     assert!(
@@ -160,10 +150,10 @@ fn missing_check_coverage_report_is_reported_as_clap_usage_error() {
 fn check_fails_fast_when_git_is_missing() {
     let temp = tempdir().expect("tempdir should exist");
 
-    let output = run_covgate_raw_with_path(
+    let output = run_covgate_raw(
         temp.path(),
-        "",
         &["check".to_string(), "missing.json".to_string()],
+        &[("PATH", "")],
     );
 
     assert_eq!(output.status.code(), Some(1));
@@ -178,7 +168,7 @@ fn check_fails_fast_when_git_is_missing() {
 fn help_lists_record_base_as_subcommand() {
     let temp = tempdir().expect("tempdir should exist");
 
-    let output = run_covgate_raw(temp.path(), &["--help".to_string()]);
+    let output = run_covgate_raw(temp.path(), &["--help".to_string()], &[]);
     assert_eq!(output.status.code(), Some(0));
     let stdout = String::from_utf8(output.stdout).expect("stdout should be utf8");
     assert!(stdout.contains("Commands:"), "stdout={stdout}");
@@ -197,7 +187,7 @@ fn version_switches_report_current_binary_version() {
     let temp = tempdir().expect("tempdir should exist");
 
     for args in [vec!["--version".to_string()], vec!["-V".to_string()]] {
-        let output = run_covgate_raw(temp.path(), &args);
+        let output = run_covgate_raw(temp.path(), &args, &[]);
         assert_eq!(output.status.code(), Some(0), "args={args:?}");
         let stdout = String::from_utf8(output.stdout).expect("stdout should be utf8");
         assert!(
@@ -211,7 +201,7 @@ fn version_switches_report_current_binary_version() {
 fn check_help_describes_arguments_and_options() {
     let temp = tempdir().expect("tempdir should exist");
 
-    let output = run_covgate_raw(temp.path(), &["check".to_string(), "--help".to_string()]);
+    let output = run_covgate_raw(temp.path(), &["check".to_string(), "--help".to_string()], &[]);
     assert_eq!(output.status.code(), Some(0));
     let stdout = String::from_utf8(output.stdout).expect("stdout should be utf8");
     assert!(stdout.contains("Arguments:"), "stdout={stdout}");
@@ -239,6 +229,7 @@ fn record_base_help_is_user_focused() {
     let output = run_covgate_raw(
         temp.path(),
         &["record-base".to_string(), "--help".to_string()],
+        &[],
     );
     assert_eq!(output.status.code(), Some(0));
     let stdout = String::from_utf8(output.stdout).expect("stdout should be utf8");
@@ -268,7 +259,7 @@ fn record_base_is_idempotent() {
     let worktree = setup_fixture_worktree(temp.path(), fixture);
     run_git(&worktree, &["branch", "-M", "task/idempotent"]);
 
-    let first = run_covgate_raw(&worktree, &["record-base".to_string()]);
+    let first = run_covgate_raw(&worktree, &["record-base".to_string()], &[]);
     assert_eq!(first.status.code(), Some(0));
     let first_ref = std::process::Command::new("git")
         .args(["rev-parse", "--verify", "refs/worktree/covgate/base"])
@@ -281,7 +272,7 @@ fn record_base_is_idempotent() {
     run_git(&worktree, &["add", "."]);
     run_git(&worktree, &["commit", "-m", "change after record-base"]);
 
-    let second = run_covgate_raw(&worktree, &["record-base".to_string()]);
+    let second = run_covgate_raw(&worktree, &["record-base".to_string()], &[]);
     assert_eq!(second.status.code(), Some(0));
     let second_stdout = String::from_utf8(second.stdout).expect("stdout should be utf8");
     assert!(second_stdout.contains("Base already recorded"));
@@ -302,7 +293,7 @@ fn record_base_refreshes_after_branch_switch() {
     let worktree = setup_fixture_worktree(temp.path(), fixture);
     run_git(&worktree, &["branch", "-M", "task/base"]);
 
-    let first = run_covgate_raw(&worktree, &["record-base".to_string()]);
+    let first = run_covgate_raw(&worktree, &["record-base".to_string()], &[]);
     assert_eq!(first.status.code(), Some(0));
     let first_ref = std::process::Command::new("git")
         .args(["rev-parse", "--verify", "refs/worktree/covgate/base"])
@@ -316,7 +307,7 @@ fn record_base_refreshes_after_branch_switch() {
     run_git(&worktree, &["add", "."]);
     run_git(&worktree, &["commit", "-m", "refresh branch work"]);
 
-    let second = run_covgate_raw(&worktree, &["record-base".to_string()]);
+    let second = run_covgate_raw(&worktree, &["record-base".to_string()], &[]);
     assert_eq!(second.status.code(), Some(0));
     let second_stdout = String::from_utf8(second.stdout).expect("stdout should be utf8");
     assert!(
@@ -355,7 +346,7 @@ fn covgate_includes_dirty_worktree_changes_by_default() {
     )
     .expect("config should be written");
 
-    let output = run_covgate(&worktree, fixture, &[]);
+    let output = run_covgate(&worktree, &fixture.coverage_json(), &[], &[]);
 
     assert_eq!(output.status.code(), Some(0));
     let stderr = String::from_utf8(output.stderr).expect("stderr should be utf8");
@@ -377,13 +368,14 @@ fn diff_file_mode_skips_dirty_worktree_guard() {
     )
     .expect("config should be written");
 
-    let output = run_covgate_with_coverage(
+    let output = run_covgate(
         &worktree,
         &fixture.coverage_json(),
         &[
             "--diff-file".to_string(),
             diff_file.to_string_lossy().into_owned(),
         ],
+        &[],
     );
 
     assert_eq!(output.status.code(), Some(0));
@@ -409,7 +401,7 @@ fn git_base_mode_errors_on_coverage_untracked_files() {
 
     run_git(&worktree, &["rm", "--cached", "src/lib.rs"]);
 
-    let output = run_covgate(&worktree, fixture, &[]);
+    let output = run_covgate(&worktree, &fixture.coverage_json(), &[], &[]);
 
     assert_eq!(output.status.code(), Some(1));
     let stderr = String::from_utf8(output.stderr).expect("stderr should be utf8");
@@ -433,7 +425,7 @@ fn git_base_mode_passes_for_uncovered_untracked_file() {
     fs::write(worktree.join("new_untracked.rs"), "pub fn pending() {}\n")
         .expect("untracked file should write");
 
-    let output = run_covgate(&worktree, fixture, &[]);
+    let output = run_covgate(&worktree, &fixture.coverage_json(), &[], &[]);
 
     assert_eq!(output.status.code(), Some(0));
     let stderr = String::from_utf8(output.stderr).expect("stderr should be utf8");
@@ -454,13 +446,14 @@ fn diff_file_mode_skips_untracked_files_check() {
     )
     .expect("config should be written");
 
-    let output = run_covgate_with_coverage(
+    let output = run_covgate(
         &worktree,
         &fixture.coverage_json(),
         &[
             "--diff-file".to_string(),
             diff_file.to_string_lossy().into_owned(),
         ],
+        &[],
     );
 
     assert_eq!(output.status.code(), Some(0));
@@ -480,7 +473,7 @@ fn automatic_base_prefers_standard_branch_ref_over_recorded_worktree_ref() {
     init_git_repo(&worktree);
     run_git(&worktree, &["branch", "-M", "task/recorded-base"]);
 
-    let output = run_covgate_raw(&worktree, &["record-base".to_string()]);
+    let output = run_covgate_raw(&worktree, &["record-base".to_string()], &[]);
     assert_eq!(output.status.code(), Some(0));
     run_git(&worktree, &["branch", "main", "HEAD"]);
 
@@ -493,7 +486,7 @@ fn automatic_base_prefers_standard_branch_ref_over_recorded_worktree_ref() {
     run_git(&worktree, &["add", "."]);
     run_git(&worktree, &["commit", "-m", "feature change"]);
 
-    let output = run_covgate(&worktree, fixture, &[]);
+    let output = run_covgate(&worktree, &fixture.coverage_json(), &[], &[]);
 
     assert_eq!(output.status.code(), Some(0));
 }
@@ -510,7 +503,7 @@ fn explicit_base_overrides_recorded_worktree_ref() {
     init_git_repo(&worktree);
     run_git(&worktree, &["branch", "-M", "task/explicit-base"]);
 
-    let output = run_covgate_raw(&worktree, &["record-base".to_string()]);
+    let output = run_covgate_raw(&worktree, &["record-base".to_string()], &[]);
     assert_eq!(output.status.code(), Some(0));
     run_git(&worktree, &["branch", "main", "HEAD"]);
 
@@ -525,8 +518,9 @@ fn explicit_base_overrides_recorded_worktree_ref() {
 
     let output = run_covgate(
         &worktree,
-        fixture,
+        &fixture.coverage_json(),
         &["--base".to_string(), "main".to_string()],
+        &[],
     );
 
     assert_eq!(output.status.code(), Some(0));
@@ -537,7 +531,7 @@ fn failure_text_requires_git_repo_when_run_outside_repository() {
     let fixture = rust_basic_pass_fixture();
     let temp = tempdir().expect("tempdir should exist");
 
-    let output = run_covgate_with_coverage(temp.path(), &fixture.coverage_json(), &[]);
+    let output = run_covgate(temp.path(), &fixture.coverage_json(), &[], &[]);
 
     assert_eq!(output.status.code(), Some(1));
     let stderr = String::from_utf8(output.stderr).expect("stderr should be utf8");
@@ -562,13 +556,14 @@ fn markdown_summary_rust_fixture() {
 
     let output = run_covgate(
         &worktree,
-        fixture,
+        &fixture.coverage_json(),
         &[
             "--diff-file".to_string(),
             diff_file.to_string_lossy().into_owned(),
             "--markdown-output".to_string(),
             markdown_output.to_string_lossy().into_owned(),
         ],
+        &[],
     );
 
     assert_eq!(output.status.code(), Some(0));
@@ -607,13 +602,14 @@ fn markdown_output_to_stdout() {
 
     let output = run_covgate(
         &worktree,
-        fixture,
+        &fixture.coverage_json(),
         &[
             "--diff-file".to_string(),
             diff_file.to_string_lossy().into_owned(),
             "--markdown-output".to_string(),
             "-".to_string(),
         ],
+        &[],
     );
 
     assert_eq!(output.status.code(), Some(0));
@@ -641,13 +637,14 @@ fn markdown_output_to_file_regression() {
 
     let output = run_covgate(
         &worktree,
-        fixture,
+        &fixture.coverage_json(),
         &[
             "--diff-file".to_string(),
             diff_file.to_string_lossy().into_owned(),
             "--markdown-output".to_string(),
             markdown_output.to_string_lossy().into_owned(),
         ],
+        &[],
     );
 
     assert_eq!(output.status.code(), Some(0));
@@ -669,7 +666,7 @@ fn github_step_summary_auto_detected() {
     )
     .expect("config should be written");
 
-    let output = run_covgate_with_env(
+    let output = run_covgate(
         &worktree,
         &fixture.coverage_json(),
         &[
@@ -701,7 +698,7 @@ fn github_step_summary_suppressed_by_flag() {
     )
     .expect("config should be written");
 
-    let output = run_covgate_with_env(
+    let output = run_covgate(
         &worktree,
         &fixture.coverage_json(),
         &[
@@ -736,7 +733,7 @@ fn github_step_summary_and_explicit_output_both_write() {
     )
     .expect("config should be written");
 
-    let output = run_covgate_with_env(
+    let output = run_covgate(
         &worktree,
         &fixture.coverage_json(),
         &[
@@ -772,13 +769,14 @@ fn markdown_output_file_write_error_is_reported() {
 
     let output = run_covgate(
         &worktree,
-        fixture,
+        &fixture.coverage_json(),
         &[
             "--diff-file".to_string(),
             diff_file.to_string_lossy().into_owned(),
             "--markdown-output".to_string(),
             temp.path().to_string_lossy().into_owned(),
         ],
+        &[],
     );
 
     assert_eq!(output.status.code(), Some(1));
@@ -805,7 +803,7 @@ fn github_step_summary_write_error_is_reported() {
     )
     .expect("config should be written");
 
-    let output = run_covgate_with_env(
+    let output = run_covgate(
         &worktree,
         &fixture.coverage_json(),
         &[
@@ -836,7 +834,7 @@ fn github_step_summary_open_error_is_reported() {
     )
     .expect("config should be written");
 
-    let output = run_covgate_with_env(
+    let output = run_covgate(
         &worktree,
         &fixture.coverage_json(),
         &[
@@ -873,11 +871,12 @@ fn path_scoped_gates_render_labeled_minimal_output() {
 
     let output = run_covgate(
         &worktree,
-        fixture,
+        &fixture.coverage_json(),
         &[
             "--diff-file".to_string(),
             diff_file.to_string_lossy().into_owned(),
         ],
+        &[],
     );
 
     assert_eq!(output.status.code(), Some(1));
@@ -906,11 +905,12 @@ fn path_scoped_gates_accept_single_string_include_and_exclude() {
 
     let output = run_covgate(
         &worktree,
-        fixture,
+        &fixture.coverage_json(),
         &[
             "--diff-file".to_string(),
             diff_file.to_string_lossy().into_owned(),
         ],
+        &[],
     );
 
     assert_eq!(output.status.code(), Some(0));
@@ -932,13 +932,14 @@ fn path_scoped_gates_markdown_adds_gate_column() {
 
     let output = run_covgate(
         &worktree,
-        fixture,
+        &fixture.coverage_json(),
         &[
             "--diff-file".to_string(),
             diff_file.to_string_lossy().into_owned(),
             "--markdown-output".to_string(),
             markdown_output.to_string_lossy().into_owned(),
         ],
+        &[],
     );
 
     assert_eq!(output.status.code(), Some(1));
@@ -970,13 +971,14 @@ fn path_scoped_gates_markdown_labels_single_named_fallback() {
 
     let output = run_covgate(
         &worktree,
-        fixture,
+        &fixture.coverage_json(),
         &[
             "--diff-file".to_string(),
             diff_file.to_string_lossy().into_owned(),
             "--markdown-output".to_string(),
             markdown_output.to_string_lossy().into_owned(),
         ],
+        &[],
     );
 
     assert_eq!(output.status.code(), Some(0));
@@ -1005,11 +1007,12 @@ fn path_scoped_gates_reject_overlap_on_changed_files() {
 
     let output = run_covgate(
         &worktree,
-        fixture,
+        &fixture.coverage_json(),
         &[
             "--diff-file".to_string(),
             diff_file.to_string_lossy().into_owned(),
         ],
+        &[],
     );
 
     assert_eq!(output.status.code(), Some(1));
@@ -1035,11 +1038,12 @@ fn path_scoped_gates_require_a_fallback_for_unmatched_supported_files() {
 
     let output = run_covgate(
         &worktree,
-        fixture,
+        &fixture.coverage_json(),
         &[
             "--diff-file".to_string(),
             diff_file.to_string_lossy().into_owned(),
         ],
+        &[],
     );
 
     assert_eq!(output.status.code(), Some(1));
@@ -1069,13 +1073,14 @@ fn absolute_llvm_paths_match_diff_fixture() {
     )
     .expect("config should be written");
 
-    let output = run_covgate_with_coverage(
+    let output = run_covgate(
         &worktree,
         &coverage_json,
         &[
             "--diff-file".to_string(),
             diff_file.to_string_lossy().into_owned(),
         ],
+        &[],
     );
 
     assert_eq!(output.status.code(), Some(0));
@@ -1107,8 +1112,9 @@ fn pr_branch_against_main_fixture() {
 
     let output = run_covgate(
         &worktree,
-        fixture,
+        &fixture.coverage_json(),
         &["--base".to_string(), "main".to_string()],
+        &[],
     );
 
     assert_eq!(output.status.code(), Some(0));
@@ -1141,7 +1147,7 @@ fn uses_repo_config_defaults_for_base_and_threshold() {
     run_git(&worktree, &["add", "covgate.toml"]);
     run_git(&worktree, &["commit", "-m", "add covgate defaults"]);
 
-    let output = run_covgate(&worktree, fixture, &[]);
+    let output = run_covgate(&worktree, &fixture.coverage_json(), &[], &[]);
 
     assert_eq!(output.status.code(), Some(0));
     let stdout = String::from_utf8(output.stdout).expect("stdout should be utf8");
@@ -1176,7 +1182,7 @@ fn uses_repo_config_defaults_from_parent_directory() {
     run_git(&worktree, &["commit", "-m", "add covgate defaults"]);
 
     let nested_dir = worktree.join("src");
-    let output = run_covgate(&nested_dir, fixture, &[]);
+    let output = run_covgate(&nested_dir, &fixture.coverage_json(), &[], &[]);
 
     assert_eq!(output.status.code(), Some(0));
     let stdout = String::from_utf8(output.stdout).expect("stdout should be utf8");
@@ -1198,13 +1204,14 @@ fn unknown_coverage_json_shape_reports_supported_formats() {
     )
     .expect("config should be written");
 
-    let output = run_covgate_with_coverage(
+    let output = run_covgate(
         &worktree,
         &invalid_coverage,
         &[
             "--diff-file".to_string(),
             diff_file.to_string_lossy().into_owned(),
         ],
+        &[],
     );
 
     assert_eq!(output.status.code(), Some(1));
@@ -1229,7 +1236,7 @@ fn minimal_pass_output_is_token_efficient() {
     )
     .expect("config should be written");
 
-    let output = run_covgate(&worktree, fixture, &[]);
+    let output = run_covgate(&worktree, &fixture.coverage_json(), &[], &[]);
 
     assert_eq!(output.status.code(), Some(0));
     let stdout = String::from_utf8(output.stdout).expect("stdout should be utf8");
@@ -1261,7 +1268,7 @@ fn minimal_fail_output_is_focused() {
     )
     .expect("config should be written");
 
-    let output = run_covgate(&worktree, fixture, &[]);
+    let output = run_covgate(&worktree, &fixture.coverage_json(), &[], &[]);
 
     assert_eq!(output.status.code(), Some(1));
     let stdout = String::from_utf8(output.stdout).expect("stdout should be utf8");
@@ -1299,13 +1306,14 @@ fn overall_coverage_remains_global_when_scoped_gates_are_configured() {
 
     let output = run_covgate(
         &worktree,
-        fixture,
+        &fixture.coverage_json(),
         &[
             "--diff-file".to_string(),
             diff_file.to_string_lossy().into_owned(),
             "--markdown-output".to_string(),
             markdown_output.to_string_lossy().into_owned(),
         ],
+        &[],
     );
 
     if !markdown_output.exists() {
